Put regex in seperate file
