+++ lib/GitHubServer.js
@@ -38,54 +38,6 @@ GitHubServer.prototype.bumpAllPullRequests = function(repo) {
     });
 };
 
-/** Find unique GoogleGroup links in body of comments (probably does not belong in this file)
- *
- * @param {String[]} comments The returned comments from GitHub.
- * @return {String[] | undefined} Unique links
- */
-GitHubServer.findGoogleGroupLinksWithRegex = function(comments) {
-    if (!defined(comments)) {
-        return;
-    }
-    var linkMatches = [];
-    for (var i = 0; i < comments.length; i++) {
-        regexSearch(comments[i], /https?:\/\/groups\.google\.com[^\s.,:]*/ig, linkMatches);
-    }
-    return linkMatches;
-};
-
-/** Find unique issue links in body of comments (probably does not belong in this file)
- *
- * @param {String[]} comments The returned comments from GitHub.
- * @return {String[] | undefined} Unique links
- */
-GitHubServer.findGitHubIssueLinksWithRegex = function(comments) {
-    if (!defined(comments)) {
-        return;
-    }
-    var linkMatches = [];
-    for (var i = 0; i < comments.length; i++) {
-        regexSearch(comments[i], /https:\/\/github\.com[^\s.,:]*/ig, linkMatches);
-    }
-    return linkMatches;
-};
-
-/** Generic regex search over block of text. Requires global regex!
- *
- * @param {String} text
- * @param {Object} regex
- * @param {String[]} matches Already-matched strings. Warning: mutated in this function.
- * @returns undefined
- */
-function regexSearch(text, regex, matches) {
-    var matchResult;
-    while ((matchResult = regex.exec(text)) !== null) {
-        if (!matches.includes(matchResult[0])) {
-            matches.push(matchResult[0]);
-        }
-    }
-}
-
 /** Send GET request to GitHub issue and return Promise to results
  *
  * @param {String} url
